Minor Like Bug and Before I Worked on Users Cont

+++ frontend/components/likes/like_button.jsx
@@ -15,20 +15,21 @@ class LikeButton extends React.Component {
     // }
 
     componentWillUnmount() { 
-        let like;
-        if (!this.state.likeId) {
-            like = { imagination_id: this.props.imagination.id, user_id: this.props.currentUser.id }
-        } else (
-            like = { id: this.state.likeId, imagination_id: this.props.imagination.id, user_id: this.props.currentUser.id }
-        )
+        if (this.props.currentUser) {
+            let like;
+            if (!this.state.likeId) {
+                like = { imagination_id: this.props.imagination.id, user_id: this.props.currentUser.id }
+            } else (
+                like = { id: this.state.likeId, imagination_id: this.props.imagination.id, user_id: this.props.currentUser.id }
+            )
 
         // debugger
-
-        if (Boolean(this.state.liked && this.props.currentUser && !this.state.likeId)) {
-            this.props.createLike(like)
-        } else if (Boolean(!this.state.liked && this.props.currentUser && this.state.likeId)) {
-            this.props.deleteLike(like)
-        } 
+            if (Boolean(this.state.liked && this.props.currentUser && !this.state.likeId)) {
+                this.props.createLike(like)
+            } else if (Boolean(!this.state.liked && this.props.currentUser && this.state.likeId)) {
+                this.props.deleteLike(like)
+            } 
+        }
     }
 
     likeOrUnlike(e) { // Right now, this only causes a re-render of the like_container, not the show itself. This is causing a lot of bugs. 
@@ -59,7 +60,15 @@ class LikeButton extends React.Component {
     render() { 
         // debugger
         return (
-            (this.state.liked) ? <button onClick={e => this.likeOrUnlike(e)}><AiFillStar />{this.props.likeCount}</button> : <button onClick={e => this.likeOrUnlike(e)}><AiOutlineStar />{this.props.likeCount}</button>
+            (this.state.liked) 
+            ? <div className="like-div" onClick={e => this.likeOrUnlike(e)}> 
+                    <div className="like-button"><AiFillStar /></div> {/* {this.props.likeCount} */}
+                <h1 className="like-text">LIKED</h1>
+              </div>
+            : <div className="like-div" onClick={e => this.likeOrUnlike(e)}>
+                <div className="like-button"><AiOutlineStar /></div> 
+                <h1 className="like-text">LIKE THIS IMAGINATION</h1>
+              </div>
             // null
         )
     }
